+++ ProjectModern/ProjectModern/Film.h
@@ -51,7 +51,7 @@ public:
 
 	std::string GetAgeRange() const;
 
-	unsigned int GetSumOfGrades() const;
+	float GetSumOfGrades() const;
 
 	// ---- SETTERS ----
 
@@ -81,14 +81,14 @@ public:
 
 	void SetAgeRange(const std::string& ageRange);
 
-	void SetSumOfGrades(const unsigned int& sumOfGrades);
+	void SetSumOfGrades(const float& sumOfGrades);
 
 
 	// ---- MEMBERS ----
 private:
 	unsigned int m_id;
 	unsigned int m_numberOfReviews;
-	unsigned int m_sumOfGrades;
+	float m_sumOfGrades;
 
 
 	float m_rating;
